Added working blank doe animation
Added code to animate the doe turning from color to cardboard.
The code uses a simple recursive function with a setTimeout wrapper.
It stores the images and positions for the various frames in arrays
(should I have a seperate function that initiliazes all the arrays?
or is it fine to have them just sitting there?).

Next step is to do the same for bird animation after all of the
color "drains" from the doe into the crack in the cardboard.  The
array of frame positions works well, but for the bird, I might be
able to just increment it by a set amount within the recursive
function?  Stuff to play around with.