Added new animation functions for bird and filling of deer
Using the previous function for the blank deer animation as
a template, made two functions, one for the animation of the
bird flying, the other for the reverse of the deer "draining"
of color.  The "fill" function is called within the bird
animation function, so it's as if the bird flying away somehow
"releases" the color back to the deer.

I also added a variable (is_active) to keep track if any
animations are currently running. This way I can make sure
that two animations don't get launched concurrently, which
would cause all sorts of problems with rendering.

Though it could be nice if the deer started to fill up as the
bird was flying away, instead of after... think of a way to
intertwine those two animations without messing everything up/
having to create a new function?