Added function to check if cursor was in deer body
Added a function that draws an invisible path around
the body of the deer on the canvas, then uses the canvas'
isPointInPath method to determine if the cursor is in the
deer.  This way, the blank deer/fill deer animation will
only be triggered when the deer itself is clicked on, not
just the ground.