Started messing around with css, possible text portion
Looking into making more of the page interactive via javascript
and css, specifically several overlapping blocks of text, where one of
them comes to the front depending on what is clicked in the image.

Also started tracing out regions for the different blocks in the fore-
ground, so that eventually the mouseover will be able to distinguish
where in the foreground the cursor is.