Started adding necessary variables to animate the sky changing
Adding an array of positions that define paths around each of the hexagonal "blocks" of corrugated cardboard that make up the sky. This will allow for different colors to be used with the hit region canvas to check what part of the sky the cursor is in, as well as allow for different animations (using clipping regions) based on where the sky was clicked, without requiring additional image files.

The site is starting to be a bit slow to load on slower connections, which is possibly because each frame of the deer "draining" of color is a separate image.  Should look into just using clipping paths for that as well, so we could use a single image of a blank deer and one of a filled deer, and the animation would lie entirely in how the clipping regions changed.