Implemented Filling animation for deer with clipping
Now that both the "drain" and "fill" animations for the deer are
done through clipping, the only files I need for these animations
are the full image and the full blank doe. I've commented out the
original functions (decided to leave them in for now in case I
want to go back), but by not having to load an additional 7 image
files, the page should load faster.

The clipping seems to work fairly well/be pretty straight forward.
Next step is to figure out how to set up rows/cols for the sky
blocks, then figure out the best way to call an expanding "hexagon"
centered on wherever the user clicks in the sky. Have some ideas,
but they still need to be fleshed out...