Started drawing sky cells to hit-region canvas
In the drawRegions function, I set up the initial draw sky-cell
portion to just draw the first row of cells, drawing each one with
a slightly different color depending on it's column.  I also changed
the getAreaName function to properly distinguish between the different
sky cells based on the underlying color on the hit-regions canvas.

It seems to work great, but still need to map out the regions for all
of the different sky cells (going to take a while as is) and figure
out how to properly grid them/call the animation.  I feel like I'm
getting close, except that the cells left of column ~5 are all fucked
up, whereas the rest of the sky is very regular...  Maybe instead of
clipping a region defined by the outside edge, I should just draw
every cell individually in that region? that seems like it would take
a lot longer, but maybe not (and might be easier to implement?)...
something to think about.