Started adding new sky-cells
Set up the first two rows of sky cell points.  Also set up a framework for
the next rows as well - basically, only the bottom three points need to be
defined, as the top three points are defined by the row above.  In
reality, I only need three points for the left-most cell of each row, then
each cell after that just needs 2 points (bottom and bottom-right), as the
bottom-left is defined by the cell to the left.  This means I still have
to manually enter two points per cell, but the majority of it should be
automated.  I think how it's set up should work, but I need to get more
points written down before really trying it out.