All sky-cells now on hit-region canvas, changed text layers
All of the sky-cells are now properly defined with 6 points, and each of
them is drawn to the hit region canvas with a distinct color. The next
step is to figure out how to measure distances to the corners of the grid
in order to calculate how many frames of animation it will take to expand
from a certain cell, and to actually do the animation. NOTE: there are
still 2 cells on the left edge that haven't been handled, need to figure
out the best way to do that...

I also changed the way that the text layers work, so that a new layer will
be brought to the foreground when your mouse hovers over something that is
interactive (it also does the ground right now as well, but I'm not sure
if I want to make that interactive?)

An easy next step might be to setup an animation for the sky just starting
from a specific cell no matter where the user clicks, to get a feel for
how the expadning set of hexagons will need to be drawn.