Implemented (semi-)working sky animation.
Sky animation is now interactive.  However, it only runs a fixed number of
frames regardless of where you click (doesn't take into account how many
frames it would take to fully convert the sky).  And, it still doesn't
handle the wonky section to the left of the deer at all, which causes the
animation to crash once it reaches that area...