Changed sky-block-regions object to array, fixed style
In an effort to try to speed up some of the code, I changed the object of
objects into an array of arrays.  Not sure if this actually sped anything
up, but it should make the size of the data objects smaller, and should
decrease the time it takes to reference a specific point (thought I'm not
sure if it was noticable).

Also removed the top 2 dummy texts, and changed the position of the text
layers be properly centered in the vertical direction with respect to the
canvas.