Added shift-click function to sky animation
Now, if you hold shift while clicking the sky, only the cell that you click will change. If doesn't keep track if a single cell is on/off though, so shift-clicking the same cell won't toggle it back and forth, just make it the opposite of what the rest of the sky is (also, even if you shift-click all of the cells in the sky, unless you regular click as well, the sky will still think it is in the state you originally started with).