Started changing sky animation to handle wonky area
Added a few checks to make the sky animaiton start to handle the wonky
section to the left of the deer.  Still doesn't work if you click in that
section, but for the most part if you click to the right, it will
correctly continue the animation through the wonky section.

I'm starting to think that the "easier" thing to do is to just change the
indices of the wonky seciton to mirror the regular sections.  The problem
is that that would require me to take out all of the crazy checks I
already put in, which might be a bigger problem than just continuing this
course.  I don't know, maybe that's a bit of a "sunk cost fallacy"
though...