Adding the foundation for transition animations
In order to add transition animations, need to change how the main images
are handled (need an array of images, and the transition would go from
image i to image i +- 1).  Also need to figure out the best way to handle
memory problems.

In addition, need to add a loading/progress bar, as the transition
animation is triggered upon loading the new image, and there needs to be
some way to tell the user if the next image is loading or if the website
is hanging up on something.